Bleach markdown output instead of only raw html
Bleaching individual raw html tags in BleachRawHtmlPostprocessor did not
work because bleach.clean() would add closing tags to elements that
already have a closing tag. This is due to the fact that opening tags
are observed in isolation from their corresponding closing tag, so
bleach thinks the closing tag is missing.

The solution provided in this commit is to bleach the full Markdown
output in a postprocessor. This has the implication that the html that
Markdown produces is also bleached, so typical elements like ul, ol, li,
pre, code, h1, strong, em, a, img, etc. need to be whitelisted. The
default whitelists have been accommodated for that.